refactor(goesr): rename netcdf_time module to dataset_time
